+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+Features:
+
+- Set the default HTTP healthcheck to `/health` and expose the `healthcheck_endpoint` variable.
+
 ## v0.1.0 (2023-07-25)
 
 Features: